Warn about possible error in #set_label_to_hashes()
The response should be ['label_name', 'label_name'], where the list
length is the same as the number of hashes, but sometimes just []
is returned. There does not appear to be a limit based on watching
the AJAX requests, and exceeding the POST max size would most
likely raise a very different error.
Basically, do not raise an exception. Just warn.